+++ ant_pos/public/antPOS/assets/emitter-CQ8_D36T.js
@@ -1,1 +1,0 @@
-import{m as t}from"./vendor-h-qvKVkq.js";const e=t();export{e};