+++ ant_pos/public/antPOS/assets/emitter-xEsCj5ku.js
@@ -1,0 +1,1 @@
+import{m as t}from"./vendor-DYmE6gU6.js";const e=t();export{e};